move nms2 to contrib, test=develop
