+++ python/paddle/fluid/tests/test_detection.py
@@ -557,8 +557,9 @@ class TestMulticlassNMS2(unittest.TestCase):
             bboxes = layers.data(
                 name='bboxes', shape=[-1, 10, 4], dtype='float32')
             scores = layers.data(name='scores', shape=[-1, 10], dtype='float32')
-            output = layers.multiclass_nms2(bboxes, scores, 0.3, 400, 200, 0.7)
-            output2, index = layers.multiclass_nms2(
+            output = fluid.contrib.multiclass_nms2(bboxes, scores, 0.3, 400,
+                                                   200, 0.7)
+            output2, index = fluid.contrib.multiclass_nms2(
                 bboxes, scores, 0.3, 400, 200, 0.7, return_index=True)
             self.assertIsNotNone(output)
             self.assertIsNotNone(output2)
